ai-ml/ocr/requirements.txt
pandas
pytesseract 
pdf2image 
Pillow 
openai
gptcache
python-dotenv
Flask